refactor(layout-kit): rename modal dir to sheet
Frees up the "modal" name for a sibling paradigm and makes this dir's
scope explicit: bottom-anchored sheets, not modals generally.
