+++ src/components/modals/upload-lesson/index.vue
@@ -5,7 +5,7 @@ import UiButton from '@/components/ui-kit/button.vue'
 import UiInput from '@/components/ui-kit/input.vue'
 import UiIcon from '@/components/ui-kit/icon.vue'
 import UiProgressBar from '@/components/ui-kit/progress-bar.vue'
-import MobileSheet from '@/components/layout-kit/modal/mobile-sheet.vue'
+import MobileSheet from '@/components/layout-kit/sheet/mobile-sheet.vue'
 import ScriptSelect from './script-select.vue'
 import { useStartLessonMutation, EdgeFunctionError } from '@/api/lessons'
 import type { LessonUploadProgress } from '@/api/lessons'
